разнес код по отдельным компонентам
