+++ src/components/exit-button/exit-button.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const ExitButton = () => <button type="button" className="player__exit">Exit</button>;
+
+export default ExitButton;
