+++ src/components/movie-card/movie-card.jsx
@@ -5,6 +5,8 @@ import VideoPlayer from '../video-player/video-player';
 import {RoutePaths} from '../../utils/constatns';
 import {idProp, previewProp, nameProp} from '../props/movie-props';
 
+const VIDEO_PREVIEW_DELAY = 1000;
+
 const MovieCardImage = ({preview, name}) => <img src={preview} alt={name} width="280" height="175" />;
 
 const MovieCardTitle = ({movieId, name}) => (
@@ -21,7 +23,7 @@ const MovieCard = ({movieId, name, preview_image: preview, onMovieCardMouseEnter
     setDelayHandler(setTimeout(() => {
       onMovieCardMouseEnter(movieId);
 
-    }, 1000));
+    }, VIDEO_PREVIEW_DELAY));
   };
 
   const handleCardMouseLeave = () => {
