+++ src/components/movies-list/movies-list.jsx
@@ -3,6 +3,10 @@ import PropTypes from 'prop-types';
 import MovieCard from '../movie-card/movie-card';
 import {movieProp} from '../props/movie-props';
 
+const getMovieCardComponent = (id, rest, onMovieCardMouseEnter, onMovieCardMouseLeave, currentMovieId) => {
+  return <MovieCard key={id} movieId={id} {...rest} onMovieCardMouseEnter={onMovieCardMouseEnter} onMovieCardMouseLeave={onMovieCardMouseLeave} currentMovieId={currentMovieId} />;
+};
+
 const MoviesList = ({allMovies}) => {
   const [currentMovieId, setCurrentMovieId] = useState(-1);
 
@@ -12,7 +16,7 @@ const MoviesList = ({allMovies}) => {
 
   return (
     <div className="catalog__movies-list">
-      {allMovies.map(({id, ...rest}) => <MovieCard key={id} movieId={id} {...rest} onMovieCardMouseEnter={onMovieCardMouseEnter} onMovieCardMouseLeave={onMovieCardMouseLeave} currentMovieId={currentMovieId} />)}
+      {allMovies.map(({id, ...rest}) => getMovieCardComponent(id, rest, onMovieCardMouseEnter, onMovieCardMouseLeave, currentMovieId))}
     </div>
   );
 };
